fix：移除无用插件
