+++ package.json
@@ -13,10 +13,8 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
-    "pinia": "^2.0.36",
     "sass": "^1.62.1",
-    "vue": "^3.3.2",
-    "vue-router": "^4.2.0"
+    "vue": "^3.3.2"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.2.0",
@@ -36,7 +34,6 @@
     "prettier": "^2.8.8",
     "typescript": "~5.0.4",
     "vite": "^4.3.5",
-    "vitest": "^0.31.0",
-    "vue-tsc": "^1.6.4"
+    "vitest": "^0.31.0"
   }
 }
