--- src/router/index.ts
@@ -1,22 +1,0 @@
-/*
- * @Author: dushuai
- * @Date: 2023-05-25 15:46:39
- * @LastEditors: dushuai
- * @LastEditTime: 2023-05-25 18:05:14
- * @description: 心平气和
- */
-import { createRouter, createWebHistory } from 'vue-router'
-import Index from '../views/Index.vue'
-
-const router = createRouter({
-  history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [
-    {
-      path: '/',
-      name: 'home',
-      component: Index
-    }
-  ]
-})
-
-export default router
